move gatsby-theme-series utils to series-utils pkg
