+++ packages/gatsby-theme-series/create-pages.js
@@ -1,4 +1,4 @@
-const withDefaults = require('./utils/default-options')
+const withDefaults = require('series-utils/default-options')
 
 const SeriesIndex = require.resolve('./src/templates/series-index.js')
 const SeriesTemplate = require.resolve('./src/templates/series.js')
