+++ packages/series-utils/default-options.js
