+++ packages/series-utils/index.js
