+++ packages/series-utils/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "series-utils",
+  "version": "1.0.0",
+  "main": "index.js",
+  "author": "Mark Chandler",
+  "license": "MIT",
+  "dependencies": {
+    "lodash": "^4.17.15"
+  }
+}
